requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
pymysql
werkzeug
pandas
cryptography
python-docx
openpyxl
pytz